Rename reportportal folder to report_portal
Folder for ReportPortal module should has report_portal name. In other
case looks like cucumber can not initialise it and throw error:
cannot load such file -- report_portal/cucumber/formatter
