+++ lib/report_portal/cucumber/formatter.rb
