+++ lib/report_portal/cucumber/json_slurper.rb
