+++ lib/report_portal/cucumber/legacy_formatter.rb
