+++ lib/report_portal/cucumber/parallel_formatter.rb
