+++ lib/report_portal/cucumber/parallel_report.rb
