+++ lib/report_portal/cucumber/report.rb
