+++ lib/report_portal/logging/log4r_outputter.rb
