+++ lib/report_portal/logging/logger.rb
