+++ lib/report_portal/logging/logging_appender.rb
