+++ lib/report_portal/patches/rest_client.rb
