+++ lib/report_portal/rspec/formatter.rb
