+++ lib/report_portal/settings.rb
