+++ lib/report_portal/tasks.rb
