+++ reportportal.gemspec
@@ -18,7 +18,7 @@
 
 lib = File.expand_path('../lib', __FILE__)
 $LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
-require 'reportportal/version'
+require 'report_portal/version'
 
 Gem::Specification.new do |s|
   s.name                   = 'reportportal'
